refactor(eval): split compiler.rs into module directory
Modularize the 1824-line compiler.rs into focused submodules for
maintainability. No behavior changes; identical test results
(366 unit tests pass).

New structure:
  compiler/
    mod.rs       -- types, compilation, evaluation logic
    helpers.rs   -- encoding, regex, pattern matching utilities
    tests.rs     -- all unit and property tests
